Rename OrbitProducerSideChannel to ProducerSideChannel
