+++ CMakeLists.txt
@@ -135,9 +135,9 @@ else()
   add_subdirectory(src/OrbitCaptureGgpClient)
   add_subdirectory(src/OrbitCaptureGgpService)
   add_subdirectory(src/OrbitProducer)
-  add_subdirectory(src/OrbitProducerSideChannel)
   add_subdirectory(src/OrbitService)
   add_subdirectory(src/OrbitTriggerCaptureVulkanLayer)
+  add_subdirectory(src/ProducerSideChannel)
 endif()
 
 if(WITH_VULKAN_LAYER)
